Moved Error, Action and AppliedAction-classes to the general package, because they are used in seval of the other packages. This resulted in some change to the visibility, so it is a bit more opened up than it was before.
