+++ src/hvl/projectparmorel/knowledge/ActionMap.java
@@ -12,6 +12,8 @@ import org.w3c.dom.Element;
 import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
 
+import hvl.projectparmorel.general.Action;
+
 class ActionMap {
 	private final String XML_NODE_NAME = "action";
 	/**
