+++ src/hvl/projectparmorel/knowledge/ErrorMap.java
@@ -10,6 +10,9 @@ import org.w3c.dom.Attr;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 import org.w3c.dom.NodeList;
+
+import hvl.projectparmorel.general.Action;
+
 import org.w3c.dom.Node;
 
 class ErrorMap {
