+++ src/hvl/projectparmorel/knowledge/HashErrorContextActionDirectory.java
@@ -7,6 +7,8 @@ import java.util.Set;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 
+import hvl.projectparmorel.general.Action;
+
 public class HashErrorContextActionDirectory implements ErrorContextActionDirectory {
 	private ErrorMap errors;
